Update ex1_11.cpp
The original answer used an 'if' statement, which at that point in the book was not covered yet- thus could not be used as a valid answer by way of how the author intended a student to answer given the tools that they had been taught up to that point. This version uses only the tools presented up to the point of the exercise.